+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add api server port to auth configmap
+
 ## [0.7.0] - 2023-07-12
 
 ## [0.6.0] - 2023-07-11